refactor: 프로젝트 정보 컴포넌트 공통화
