+++ src/app/(project)/project-management/[id]/page.tsx
@@ -3,6 +3,7 @@
 import BottomButton from '@/components/common/bottom-button';
 import Guide from '@/components/common/guide';
 import Icon from '@/components/common/icon';
+import ProjectInfo from '@/components/project/project-info';
 import ProjectProgress from '@/components/project/project-progress';
 import { HostProjectGuide } from '@/constants/guide';
 import useDisclosure from '@/hooks/useDisclosure';
@@ -42,34 +43,6 @@ const ProjectManagementDetailPage = () => {
           guides={HostProjectGuide.cancellation}
           collapsible
         />
-      </div>
-    </div>
-  );
-};
-
-const ProjectInfo = ({ project }: { project: Omit<IProject, 'state'> }) => {
-  return (
-    <div className="flex flex-col gap-1 rounded-[8px] bg-gray-90 px-[12px] py-[14px]">
-      <h2 className="font-title-18 text-gray-10">프로젝트</h2>
-      <h2 className="font-body-14m text-gray-10">{project.title}</h2>
-      <div className="flex items-center">
-        <p
-          className={cn(
-            'font-tag-14 text-gray-40 after:mx-[6px] after:content-["|"]',
-          )}
-        >
-          {project.location}
-        </p>
-        <p
-          className={cn(
-            'font-tag-14 text-gray-40 after:mx-[6px] after:content-["|"]',
-          )}
-        >
-          {project.date}
-        </p>
-        <p className={cn('font-tag-14 text-gray-40 after:mx-[6px]')}>
-          {project.time}
-        </p>
       </div>
     </div>
   );
